test(#32): Update PostServiceImpl
- Get the shortValue from PostStatus values provided
- Update UTs

+++ common/src/main/java/org/wcdevs/blog/core/common/post/PostServiceImpl.java
@@ -24,7 +24,7 @@ public class PostServiceImpl implements PostService {
   @Override
   @Transactional(readOnly = true)
   public Page<PostDto> getPosts(PostStatus status, Pageable pageable) {
-    return postRepository.getPosts(status, pageable);
+    return postRepository.getPosts(status.shortValue(), pageable);
   }
 
   @Override
